requirements.txt
streamlit
transformers
tqdm
arxiv
sentence-transformers
python-dotenv
huggingface-hub